Update Readme.md
This worked back in the days of Snow Leopard and then things started to change in PathFinder and then Lion, so I'm not sure if this still works. Maybe the PF preferences really will remember these keybindings! VERY WELL COULD BE, LIKELY IS, OBSOLETE, BUT I'M NOT POSITIVE, YET I'M KINDA REAL SURE.